+++ knip.config.ts
@@ -1,0 +1,7 @@
+const config = {
+  ignore: ['./sanity.cli.ts', './src/components/ui/**/*'],
+  ignoreBinaries: ['sync-dotenv', 'only-allow'],
+  ignoreDependencies: ['@commitlint/config-conventional', 'lucide-react'],
+};
+
+export default config;