+++ src/lib/utils.ts
@@ -1,6 +1,25 @@
+import { LangSchema } from '@/services/sanity/parser';
 import { type ClassValue, clsx } from 'clsx';
 import { twMerge } from 'tailwind-merge';
 
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs));
+}
+
+export function formatDateByLang(date: string, lang: LangSchema) {
+  // Check if the language is Spanish (Latin America)
+  if (lang === 'es') {
+    return new Intl.DateTimeFormat('es-ES', {
+      year: 'numeric',
+      month: 'long',
+      day: 'numeric',
+    }).format(new Date(date));
+  }
+
+  // Default to US English if no language is specified
+  return new Intl.DateTimeFormat('en-US', {
+    year: 'numeric',
+    month: 'long',
+    day: 'numeric',
+  }).format(new Date(date));
 }