+++ src/services/sanity/queries.ts
@@ -8,13 +8,46 @@ export const getLayoutQuery = groq`
 
 export const getHomeQuery = groq`
     *[_type == "home" && language == $lang][0]{
-        ...
+        ...,
+        sections[]{
+            ...,
+            _type == 'featuredArticles' => {
+                ...,
+                title,
+                items[]{
+                    ...,
+                    ...@-> {
+                        title,
+                        summary,
+                        'updateAt': _updatedAt,
+                        'slug': slug.current,
+                        estimatedReadingTime{
+                            ...
+                        }
+                    }
+                }
+            }
+        }
     }
 `;
 
 export const getLogbookQuery = groq`
     *[_type == "logbook" && language == $lang][0]{
         ...
+    }
+`;
+
+export const getBlogArticleQuery = groq`
+    *[_type == "article" && language == $lang && slug.current == $slug][0]{
+        ...,
+        'updateAt': _updatedAt
+    }
+`;
+
+export const getBlogArticlesForSitemapQuery = groq`
+    *[_type == "article"]{
+        'lang': language,
+        'slug': slug.current,
     }
 `;